+++ src-electron/zcl/zcl-loader-silabs.js
@@ -40,6 +40,7 @@ const _ = require('lodash')
 const querySessionNotification = require('../db/query-session-notification')
 const queryPackageNotification = require('../db/query-package-notification')
 const newDataModel = require('./zcl-loader-new-data-model')
+const conformParser = require('../conformance/conform-xml-parser')
 
 /**
  * Promises to read the JSON file and resolve all the data.
@@ -492,7 +493,7 @@ function prepareCluster(cluster, context, isExtension = false) {
         description: command.description ? command.description[0].trim() : '',
         source: command.$.source,
         isOptional: command.$.optional == 'true' ? true : false,
-        conformance: parseConformanceFromXML(command),
+        conformance: conformParser.parseConformanceFromXML(command),
         mustUseTimedInvoke: command.$.mustUseTimedInvoke == 'true',
         introducedIn: command.$.introducedIn,
         removedIn: command.$.removedIn,
@@ -548,7 +549,7 @@ function prepareCluster(cluster, context, isExtension = false) {
         manufacturerCode: event.$.manufacturerCode,
         name: event.$.name,
         side: event.$.side,
-        conformance: parseConformanceFromXML(event),
+        conformance: conformParser.parseConformanceFromXML(event),
         priority: event.$.priority,
         description: event.description ? event.description[0].trim() : '',
         isOptional: event.$.optional == 'true',
@@ -649,7 +650,7 @@ function prepareCluster(cluster, context, isExtension = false) {
             : attribute.$.type,
         side: attribute.$.side,
         define: attribute.$.define,
-        conformance: parseConformanceFromXML(attribute),
+        conformance: conformParser.parseConformanceFromXML(attribute),
         min: attribute.$.min,
         max: attribute.$.max,
         minLength: 0,
@@ -737,7 +738,7 @@ function prepareCluster(cluster, context, isExtension = false) {
         bit: feature.$.bit,
         defaultValue: feature.$.default,
         description: feature.$.summary,
-        conformance: parseConformanceFromXML(feature)
+        conformance: conformParser.parseConformanceFromXML(feature)
       }
 
       ret.features.push(f)
@@ -1693,7 +1694,7 @@ function prepareDeviceType(deviceType) {
             include.features[0].feature.forEach((f) => {
               features.push({
                 code: f.$.code,
-                conformance: parseConformanceFromXML(f)
+                conformance: conformParser.parseConformanceFromXML(f)
               })
             })
           }
@@ -2391,140 +2392,6 @@ async function parseFeatureFlags(db, packageId, featureFlags) {
       )
     })
   )
-}
-
-/**
- * Parses conformance from XML data.
- * The conformance could come from features, attributes, commands, or events
- *
- * Call recursive helper function to parse conformance only if the conformance exists.
- * Otherwise, return empty string directly
- *
- * An example of parsing the conformance of 'User' device type feature:
- *
- * Input operand from xml data:
- * {
- *   "$": {"code": "USR", "name": "User"},
- *   "mandatoryConform": [
- *      { "andTerm": [
- *           {
- *             "condition": [{"$": {"name": "Matter"}}],
- *             "orTerm": [
- *                 { "feature": [
- *                      { "$": {"name": "PIN"}},
- *                      { "$": {"name": "RID"}},
- *                      { "$": {"name": "FGP"}},
- *                      { "$": {"name": "FACE"}}
- *                   ]
- *                 }
- *               ]
- *            }
- *          ]
- *        }
- *    ]
- * }
- *
- * Output conformance string:
- *  "Matter & (PIN | RID | FGP | FACE)"
- *
- * @param {*} operand
- * @returns The conformance string
- */
-function parseConformanceFromXML(operand) {
-  let hasConformance = Object.keys(operand).some((key) =>
-    key.includes('Conform')
-  )
-  return hasConformance ? parseConformanceRecursively(operand) : ''
-}
-
-/**
- * helper function to parse conformance or an operand in conformance recursively
- *
- * The baseLevelTerms variable include terms that can not have nested terms.
- * When they appear, stop recursing and return the name inside directly
- *
- * @param {*} operand
- * @param {*} depth
- * @param {*} parentJoinChar
- * @returns The conformance string.
- */
-function parseConformanceRecursively(operand, depth = 0, parentJoinChar = '') {
-  if (depth > 200) {
-    throw new Error(`Maximum recursion depth exceeded
-      when parsing conformance: ${JSON.stringify(operand)}`)
-  }
-  const baseLevelTerms = ['feature', 'condition', 'attribute', 'command']
-  if (operand.mandatoryConform) {
-    let insideTerm = operand.mandatoryConform[0]
-    // Recurse further if insideTerm is not empty
-    if (insideTerm && Object.keys(insideTerm).toString() != '$') {
-      return parseConformanceRecursively(operand.mandatoryConform[0], depth + 1)
-    } else {
-      return 'M'
-    }
-  } else if (operand.optionalConform) {
-    let insideTerm = operand.optionalConform[0]
-    // check '$' key is not the only key in the object to handle special cases
-    // e.g. '<optionalConform choice="a" more="true"/>'
-    if (insideTerm && Object.keys(insideTerm).toString() != '$') {
-      return `[${parseConformanceRecursively(operand.optionalConform[0], depth + 1)}]`
-    } else {
-      return 'O'
-    }
-  } else if (operand.otherwiseConform) {
-    return Object.entries(operand.otherwiseConform[0])
-      .map(([key, value]) =>
-        parseConformanceRecursively({ [key]: value }, depth + 1)
-      )
-      .join(', ')
-  } else if (operand.notTerm) {
-    // need to surround terms inside a notTerm with '()' if it contains multiple terms
-    // e.g. !(A | B) or !(A & B)
-    // able to process multiple parallel notTerms, e.g. !A & !B
-    return operand.notTerm
-      .map((term) => {
-        let nt = parseConformanceRecursively(term, depth + 1)
-        return nt.includes('&') || nt.includes('|') ? `!(${nt})` : `!${nt}`
-      })
-      .join(` ${parentJoinChar} `)
-  } else if (operand.andTerm || operand.orTerm) {
-    // process andTerm and orTerm in the same logic
-    // when joining multiple orTerms inside andTerms, we need to
-    // surround them with '()', vice versa for andTerms inside orTerms
-    // e.g. A & (B | C) or A | (B & C)
-    let joinChar = operand.andTerm ? '&' : '|'
-    let termKey = operand.andTerm ? 'andTerm' : 'orTerm'
-    let oppositeChar = joinChar == '&' ? '|' : '&'
-    return Object.entries(operand[termKey][0])
-      .map(([key, value]) => {
-        if (baseLevelTerms.includes(key)) {
-          return value.map((operand) => operand.$.name).join(` ${joinChar} `)
-        } else {
-          let terms = parseConformanceRecursively(
-            { [key]: value },
-            depth + 1,
-            joinChar
-          )
-          return terms.includes(oppositeChar) ? `(${terms})` : terms
-        }
-      })
-      .join(` ${joinChar} `)
-  } else if (operand.provisionalConform) {
-    return 'P'
-  } else if (operand.disallowConform) {
-    return 'X'
-  } else if (operand.deprecateConform) {
-    return 'D'
-  } else {
-    // reach base level terms, return the name directly
-    for (const term of baseLevelTerms) {
-      if (operand[term]) {
-        return operand[term][0].$.name
-      }
-    }
-    // reaching here means the term is too complex to parse
-    return 'desc'
-  }
 }
 
 /**
